+++ config/config.example.js
@@ -8,7 +8,8 @@ module.exports = {
   boards: [
     {
       shortcut: 'hw',
-      board: 'hello-world',
+      name: 'hello-world',
+      id: 'board-1', // optional: only required when board is public but user should see it
       estimates_with_round_brackets: true,
       estimates_with_square_brackets: true,
     },