+++ src/__mocks__/mocks.js
@@ -2,7 +2,8 @@ import { pick } from 'lodash'
 
 // This file contains mocked responses from the data/trello API.
 export const mockExampleBoardConfig = {
-  board: 'hello-world',
+  name: 'hello-world',
+  id: 'board-1', // can be optional
   estimates_with_round_brackets: true,
   estimates_with_square_brackets: true,
   shortcut: 'hw',
@@ -47,7 +48,7 @@ export const mockExampleBoardResponse = { name: 'hello-world', id: 'board-1' }
 
 // most likely it should be an array with two lists (one matching and one not)
 export const mockExampleListResponse = {
-  id: 'list-123',
+  id: 'list-1',
   idBoard: 'board-1',
   name: 'Test List Name #upcoming',
 }