+++ src/actions/boards/__tests__/index.test.js
@@ -67,6 +67,7 @@ describe('actions/board:async actions', () => {
         error: null,
         payload: [
           {
+            idx: 0,
             board,
             config,
           },