+++ src/middleware/history.js
@@ -4,8 +4,13 @@ import { actions as appActions } from '../actions/app'
 export default store => next => action => {
   switch (action.type) {
     case appActions.TOGGLE_PREFERRED_LIST:
-      // NOTE: due to the look of the pattern similar to /pattern/ we do not have to add a / at the beginning
-      store.dispatch(push(`/pattern${action.pattern}`))
+      const { pattern } = action
+      if (!pattern) {
+        return next(action)
+      }
+
+      const path = action.pattern.charAt(0) === '/' ? action.pattern : `/${action.pattern}`
+      store.dispatch(push(`/pattern${path}`))
       return next(action)
     default:
       return next(action)